fix casting bug

+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/ui/detail/event/EventListDivider.java
@@ -43,7 +43,7 @@ public class EventListDivider extends RecyclerView.ItemDecoration {
             view.getLeft(),
             view.getBottom(),
             view.getRight(),
-            view.getBottom() + mHeightDp,
+            view.getBottom() + (float) mHeightDp,
             mPaint);
       }
     }
